# src/main/java/com/bank/credit_card/balances/controller/BalancesApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.credit_card.balances.controller;

import com.bank.credit_card.balances.schema.response.BalanceResponse;
import com.bank.credit_card.generic.schema.response.DefaultResponse;
import com.bank.credit_card.generic.schema.response.DefaultResponse4xx;
import com.bank.credit_card.generic.schema.response.Tracking;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-15T00:33:17.244211400-05:00[America/Lima]")
@Validated
@Tag(name = "balances", description = "the balances API")
public interface BalancesApi {

    /**
     * POST /cards/{idCard}/changeBalance : changeBalance
     *
     * @param idCard (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "changeBalance",
            summary = "changeBalance",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Tracking.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/cards/{idCard}/changeBalance",
            produces = {"application/json"}
    )
    ResponseEntity<Tracking> changeBalance(
            @Parameter(name = "idCard", description = "", required = true, in = ParameterIn.PATH) @PathVariable("idCard") Long idCard
    );


    /**
     * GET /cards/{cardId}/balances : getBalance
     *
     * @param cardId (required)
     * @return (status code 200)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "getBalance",
            summary = "getBalance",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "200", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = BalanceResponse.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.GET,
            value = "/cards/{cardId}/balances",
            produces = {"application/json"}
    )
    ResponseEntity<BalanceResponse> getBalance(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") String cardId
    );

}