# src/main/java/com/bank/credit_card/consumptions/controller/ConsumptionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.3.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.bank.credit_card.consumptions.controller;

import com.bank.credit_card.consumptions.schema.request.ConsumptionRequest;
import com.bank.credit_card.consumptions.schema.request.SplitConsumptionDebtRequest;
import com.bank.credit_card.generic.schema.response.DefaultResponse;
import com.bank.credit_card.generic.schema.response.DefaultResponse4xx;
import com.bank.credit_card.generic.schema.response.Tracking;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.Valid;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.BindingResult;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-02-15T00:33:17.244211400-05:00[America/Lima]")
@Validated
@Tag(name = "consumptions", description = "the consumptions API")
public interface ConsumptionsApi {

    /**
     * DELETE /consumptions/{consumptionId} : closeConsumption
     *
     * @param consumptionId (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "closeConsumption",
            summary = "closeConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Tracking.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.DELETE,
            value = "/consumptions/{consumptionId}",
            produces = {"application/json"}
    )
    ResponseEntity<Tracking> closeConsumption(
            @Parameter(name = "consumptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("consumptionId") Long consumptionId
    );


    /**
     * POST /consumptions/{consumptionId}/split : splitConsumptionDebt
     *
     * @param consumptionId               (required)
     * @param splitConsumptionDebtRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "splitConsumptionDebt",
            summary = "splitConsumptionDebt",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Tracking.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/consumptions/{consumptionId}/split",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<Tracking> splitConsumptionDebt(
            @Parameter(name = "consumptionId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("consumptionId") Long consumptionId,
            @Parameter(name = "SplitConsumptionDebtRequest", description = "", required = true) @Valid @RequestBody SplitConsumptionDebtRequest splitConsumptionDebtRequest,
            BindingResult bindingResult
    );


    /**
     * POST /cards/{cardId}/consumption : createConsumption
     *
     * @param cardId             (required)
     * @param consumptionRequest (required)
     * @return (status code 202)
     * or  (status code 400)
     * or  (status code 404)
     * or  (status code 500)
     */
    @Operation(
            operationId = "createConsumption",
            summary = "createConsumption",
            description = "",
            tags = {},
            responses = {
                    @ApiResponse(responseCode = "202", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = Tracking.class))
                    }),
                    @ApiResponse(responseCode = "400", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse4xx.class))
                    }),
                    @ApiResponse(responseCode = "404", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    }),
                    @ApiResponse(responseCode = "500", description = "", content = {
                            @Content(mediaType = "application/json", schema = @Schema(implementation = DefaultResponse.class))
                    })
            }
    )
    @RequestMapping(
            method = RequestMethod.POST,
            value = "/cards/{cardId}/consumption",
            produces = {"application/json"},
            consumes = {"application/json"}
    )
    ResponseEntity<Tracking> createConsumption(
            @Parameter(name = "cardId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("cardId") Long cardId,
            @Parameter(name = "ConsumptionRequest", description = "", required = true) @Valid @RequestBody ConsumptionRequest consumptionRequest,
            BindingResult bindingResult
    );

}