# src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  title: credit_card
  description: ''
  version: 1.0.0
...
paths:
...
  /consumptions/{consumptionId}:
    delete:
      summary: closeConsumption
      deprecated: false
      description: ''
      operationId: closeConsumption
      tags: []
      parameters:
        - name: consumptionId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /cards/{cardId}/consumption:
    post:
      summary: createConsumption
      deprecated: false
      description: ''
      operationId: createConsumption
      tags: []
      parameters:
        - name: cardId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsumptionRequest'
            examples: {}
        required: true
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /consumptions/{consumptionId}/split:
    post:
      summary: splitConsumptionDebt
      deprecated: false
      description: ''
      operationId: splitConsumptionDebt
      tags: []
      parameters:
        - name: consumptionId
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                quantity:
                  type: integer
                  minimum: 1
                  maximum: 36
              required:
                - quantity
            examples: {}
        required: true
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /cards/{cardId}/balances:
    get:
      summary: getBalance
      deprecated: false
      description: ''
      operationId: getBalance
      tags: []
      parameters:
        - name: cardId
          in: path
          description: ''
          required: true
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
  /cards/{idCard}/changeBalance:
    post:
      summary: changeBalance
      deprecated: false
      description: ''
      operationId: changeBalance
      tags: []
      parameters:
        - name: idCard
          in: path
          description: ''
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '202':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tracking'
          headers: {}
        '400':
          $ref: '#/components/responses/BadRequest'
          description: ''
        '404':
          $ref: '#/components/responses/NotFound'
          description: ''
        '500':
          $ref: '#/components/responses/InternalError'
          description: ''
      security: []
...
components:
  schemas:
...
    ConsumptionRequest:
      type: object
      properties:
        sellerName:
          type: string
        currency:
          type: string
          pattern: ^(PEN|USD)$
        amount:
          type: number
      required:
        - sellerName
        - currency
        - amount
...
    BalanceDataResponse:
      type: object
      properties:
        originName:
          type: string
        amount:
          type: number
        date:
          type: string
          format: date-time
        status:
          type: string
        typeTransaccion:
          type: string
        currency:
          type: string
        origin:
          type: string
        category:
          type: string
      required:
        - originName
        - amount
        - date
        - status
        - typeTransaccion
        - currency
    BalanceResponse:
      type: object
      properties:
        cardId:
          type: integer
          format: int64
        consumptionAmount:
          type: number
        paymentAmount:
          type: number
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date-time
        exchangeRate:
          type: number
        balanceDataset:
          type: array
          items:
            $ref: '#/components/schemas/BalanceDataResponse'
      required:
        - cardId
        - consumptionAmount
        - paymentAmount
        - startDate
        - endDate
        - balanceDataset
        - exchangeRate
    Tracking:
      type: object
      properties:
        trackingId:
          type: string
          format: uuid
        operationDate:
          type: string
          format: date-time
      required:
        - trackingId
        - operationDate
    DefaultResponse4xx:
      type: object
      properties:
        description:
          type: string
        errors:
          type: array
          items:
            type: string
      required:
        - description
        - errors
    DefaultResponse:
      type: object
      properties:
        description:
          type: string
      required:
        - description
  responses:
    BadRequest:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse4xx'
    NotFound:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse'
    InternalError:
      description: ''
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DefaultResponse'
  securitySchemes: {}
servers: []
security: []